separated services to volume and metadata
