+++ intern/resource/cv/resource.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from intern.resource import Resource
-from cloudvolume import CloudVolume
+from cloudvolume import CloudVolume, Vec
 
 import numpy as np
 from os import path
@@ -37,7 +37,7 @@ class CloudVolumeResource(Resource):
                     returned by multiprocessing.cpu_count(). When parallel > 1, shared
                     memory (Linux) or emulated shared memory via files (other platforms) 
                     is used by the underlying download.
-            **params () : keyword-value arguments for info object
+            **params () : keyword-value arguments for info object --> dict representing single mip level thats JSON encodable 
 
         Returns:
             CloudVolume : cloudvolume instance with specified parameters 
@@ -45,14 +45,20 @@ class CloudVolumeResource(Resource):
         Resource.__init__(self)
         if new_layer:
             info = CloudVolume.create_new_info(
-                num_channels = params.get('num_channels', 1),
-                layer_type = params.get('layer_type', None), # 'image' or 'segmentation'
-                data_type = params.get('data_type', 'uint8'), # can pick any popular uint, defaults to unit8
-                encoding = params.get('encoding', None), # other options: 'jpeg', 'compressed_segmentation' (req. uint32 or uint64)
-                resolution = params.get('resolution', [1,1,1]), # X,Y,Z values in nanometers
-                voxel_offset = params.get('voxel_offset', [0,0,0]), # offset in X,Y,Z voxels
-                chunk_size = params.get('chunk_size', None), # rechunk of image X,Y,Z in voxels
-                volume_size = params.get('volume_size', [1,1,1]) # X,Y,Z size in voxels
+                num_channels = params.get('num_channels'), # (int) 1 for grayscale, 3 for RGB 
+                layer_type = params.get('layer_type'), # (str)'image' or 'segmentation'
+                data_type = params.get('data_type'), # (str) e.g. 'uint8', 'uint16', 'uint32', 'float32'
+                encoding = params.get('encoding'), # (str) other options: 'jpeg', 'compressed_segmentation' (req. uint32 or uint64)
+                resolution = params.get('resolution'), # (x,y,z) X,Y,Z voxel dimensions in nanometers
+                voxel_offset = params.get('voxel_offset'), # (x,y,z) voxel offset
+                volume_size = params.get('volume_size'), # (x,y,z) extent of dataset from voxel offset
+                chunk_size = params.get('chunk_size', (64,64,64)), # rechunk of image X,Y,Z in voxels
+                mesh = params.get('mesh'), # (str) name of mesh directory, typically "mesh"
+                skeletons = params.get('skeletons'), # (str) name of skeletons directory, typically "skeletons"
+                compressed_segmentation_block_size = params.get('compressed_segmentation_block_size', (8,8,8)), #dimensions of each compressed sub-block (only used when encoding is 'compressed_segmentation')
+                max_mip = params.get('max_mip', 0), #(int)the maximum mip level id
+                factor = params.get('factor', Vec(2,2,1)), # (Vec)the downsampling factor for each mip level
+                redirect = params.get('redirect') #(str)If this volume has moved, you can set an automatic redirect by specifying a cloudpath here.
             )
             owners = params.get('owners', []) # list of contact email addresses
             description = params.get('description', 'No description provided')
@@ -77,7 +83,6 @@ class CloudVolumeResource(Resource):
         vol.commit_info() # generates protocol://bucket/dataset/layer/info json file
         self.cloudvolume = vol
 
-
     def valid_volume(self):
         """Returns True if resource is something that can access the volume service.
         Args:
@@ -85,38 +90,3 @@ class CloudVolumeResource(Resource):
             (bool) : True if calls to volume service may be made.
         """
         return True
-
-    def create_cutout(self, data, x_range, y_range, z_range):
-        """
-            Method to upload a cutout of data
-            Args:
-                data (str) : Path to the data
-                vol (CloudVolume) : Existing cloudvolume instance 
-                x_range (list) : x range within the 3D space
-                y_range (list) : y range within the 3D space
-                z_range (list) : z range witinn the 3D space
-            Retruns:
-                message (str) : Uploading Data... message
-        """
-        if x_range==[] and y_range==[] and z_range==[]:
-            self.cloudvolume[:,:,:] = data
-        else:
-            self.cloudvolume[x_range[0]:x_range[1], y_range[0]:y_range[1], z_range[0]:z_range[1]] = data
-        print("Data uploaded.")
-
-    def get_cutout(self, x_range, y_range, z_range):
-        """
-            Method to download a cutout of data
-            Args:
-                vol (CloudVolume) : Existing non-empty cloudvolume instance 
-                x_range (list) : x range within the 3D space
-                y_range (list) : y range within the 3D space
-                z_range (list) : z range within the 3D space
-            Retruns:
-                data (numpy array) : image stack from the cloud or local system
-        """
-        if x_range== [] and y_range== [] and z_range == []:
-            data = self.cloudvolume[:,:,:]
-        else:
-            data = self.cloudvolume[x_range[0]:x_range[1], y_range[0]:y_range[1], z_range[0]:z_range[1]]
-        return data
